runner refactoring done - moved every agent business logic or runner business logic to own class and access them via provider classes (untested)
