+++ src/agents/setup/__init__.py
@@ -1,0 +1,9 @@
+from .setup_agent import SetupAgent
+from .setup_agent_llm import SetupAgentLLM
+from .setup_agent_no_llm import SetupAgentNoLLM
+
+__all__ = [
+    'SetupAgent',
+    'SetupAgentLLM',
+    'SetupAgentNoLLM',
+]
