+++ src/agents/setup/setup_agent_llm.py
@@ -10,62 +10,12 @@ Analyzes 15m timeframe data and produces semantic analysis:
 
 from typing import Dict, Optional
 from src.llm import create_client, LLMConfig
-from src.config import config
+from src.config import Config
 from src.utils.logger import log
 
+from .setup_agent import SetupAgent
 
-def _compute_setup_signals(data: Dict) -> Dict[str, Optional[float]]:
-    kdj_j = data.get('kdj_j', 50)
-    trend = data.get('trend_direction', 'neutral')
-    close = data.get('close_15m', 0)
-    bb_middle = data.get('bb_middle', 0)
-    macd_diff = data.get('macd_diff', 0)
-
-    if trend == 'long':
-        if kdj_j < 40:
-            stance = 'PULLBACK_ZONE'
-            zone = 'GOOD_ENTRY'
-        elif kdj_j > 80:
-            stance = 'OVERBOUGHT'
-            zone = 'WAIT'
-        else:
-            stance = 'NEUTRAL'
-            zone = 'MONITOR'
-    elif trend == 'short':
-        if kdj_j > 60:
-            stance = 'RALLY_ZONE'
-            zone = 'GOOD_ENTRY'
-        elif kdj_j < 20:
-            stance = 'OVERSOLD'
-            zone = 'WAIT'
-        else:
-            stance = 'NEUTRAL'
-            zone = 'MONITOR'
-    else:
-        stance = 'NEUTRAL'
-        zone = 'WAIT'
-
-    if macd_diff > 0:
-        macd_signal = 'BULLISH'
-    elif macd_diff < 0:
-        macd_signal = 'BEARISH'
-    else:
-        macd_signal = 'NEUTRAL'
-
-    bb_position = 'ABOVE_MID' if close > bb_middle else 'BELOW_MID'
-
-    return {
-        'stance': stance,
-        'zone': zone,
-        'kdj_j': kdj_j,
-        'trend': trend,
-        'bb_position': bb_position,
-        'macd_signal': macd_signal,
-        'macd_diff': macd_diff
-    }
-
-
-class SetupAgentLLM:
+class SetupAgentLLM(SetupAgent):
     """
     15m Setup Analysis Agent
     
@@ -73,7 +23,10 @@ class SetupAgentLLM:
     Output: Semantic analysis paragraph
     """
     
-    def __init__(self):
+    def __init__(
+        self,
+        config: Config
+    ):
         """Initialize SetupAgentLLM with LLM client"""
         llm_config = config.llm
         provider = llm_config.get('provider', 'deepseek')
@@ -124,13 +77,13 @@ class SetupAgentLLM:
             
             # Use unified LLM interface
             response = self.client.chat(
-                system_prompt=self._get_system_prompt(),
+                system_prompt=self.get_system_prompt(),
                 user_prompt=prompt
             )
             
             analysis = response.content.strip()
             
-            signals = _compute_setup_signals(data)
+            signals = self._compute_setup_signals(data)
             
             result = {
                 'analysis': analysis,
@@ -172,7 +125,7 @@ class SetupAgentLLM:
                 'metadata': {'error': str(e)}
             }
     
-    def _get_system_prompt(self) -> str:
+    def get_system_prompt(self) -> Optional[str]:
         """System prompt for setup analysis"""
         return """You are a professional crypto setup analyst. Your task is to analyze 15m timeframe data and assess entry positions using KDJ, Bollinger Bands, and MACD.
 
@@ -271,70 +224,3 @@ Consider: For LONG, we want pullback (KDJ<40 or near lower BB) + bullish MACD. F
                 return f"15m is oversold with KDJ_J={kdj_j:.0f}. Wait for rally before entering short positions."
             else:
                 return f"15m is in neutral zone with KDJ_J={kdj_j:.0f}. Wait for clearer rally signal."
-
-
-class SetupAgent:
-    """
-    15m Setup Analysis Agent (no LLM)
-
-    Uses rule-based heuristics only.
-    """
-
-    def __init__(self):
-        log.info("📊 Setup Agent (no LLM) initialized")
-
-    def analyze(self, data: Dict) -> Dict:
-        signals = _compute_setup_signals(data)
-        analysis = self._get_fallback_analysis(data)
-        result = {
-            'analysis': analysis,
-            'stance': signals['stance'],
-            'metadata': {
-                'zone': signals['zone'],
-                'kdj_j': round(signals['kdj_j'], 1),
-                'trend': signals['trend'].upper(),
-                'bb_position': signals['bb_position'],
-                'macd_signal': signals['macd_signal'],
-                'macd_diff': round(signals['macd_diff'], 2)
-            }
-        }
-        log.info(f"📊 Setup Agent (no LLM) [{signals['stance']}] (Zone: {signals['zone']}, KDJ: {signals['kdj_j']:.1f}) for {data.get('symbol', 'UNKNOWN')}")
-
-        try:
-            from src.server.state import global_state
-            if hasattr(global_state, 'saver') and hasattr(global_state, 'current_cycle_id'):
-                global_state.saver.save_setup_analysis(
-                    analysis=analysis,
-                    input_data=data,
-                    symbol=data.get('symbol', 'UNKNOWN'),
-                    cycle_id=global_state.current_cycle_id,
-                    model='rule_based'
-                )
-        except Exception as e:
-            log.warning(f"Failed to save setup analysis log: {e}")
-
-        return result
-
-    def _get_fallback_analysis(self, data: Dict) -> str:
-        """Fallback analysis using rule-based heuristics"""
-        kdj_j = data.get('kdj_j', 50)
-        trend = data.get('trend_direction', 'neutral')
-        close = data.get('close_15m', 0)
-        bb_middle = data.get('bb_middle', 0)
-        
-        if trend == 'long':
-            if kdj_j < 40:
-                return f"15m setup shows pullback zone with KDJ_J={kdj_j:.0f}. Good entry area for long positions. Price is {'below' if close < bb_middle else 'above'} BB middle."
-            elif kdj_j > 80:
-                return f"15m is overbought with KDJ_J={kdj_j:.0f}. Wait for pullback before entering long positions."
-            else:
-                return f"15m is in neutral zone with KDJ_J={kdj_j:.0f}. Wait for clearer pullback signal."
-        elif trend == 'short':
-            if kdj_j > 60:
-                return f"15m setup shows rally zone with KDJ_J={kdj_j:.0f}. Good entry area for short positions."
-            elif kdj_j < 20:
-                return f"15m is oversold with KDJ_J={kdj_j:.0f}. Wait for rally before entering short positions."
-            else:
-                return f"15m is in neutral zone with KDJ_J={kdj_j:.0f}. Wait for clearer rally signal."
-        else:
-            return f"15m shows neutral setup with KDJ_J={kdj_j:.0f}. No clear entry signal."
