+++ src/agents/trend/__init__.py
@@ -1,0 +1,9 @@
+from .trend_agent import TrendAgent
+from .trend_agent_llm import TrendAgentLLM
+from .trent_agent_no_llm import TrendAgentNoLLM
+
+__all__ = [
+    'TrendAgent',
+    'TrendAgentLLM',
+    'TrendAgentNoLLM',
+]
